Refactorizado modificar sección
