+++ frontend/views/casas/_crear-seccion.php
@@ -1,57 +1,7 @@
 <?php
 
-use yii\helpers\Url;
-use yii\helpers\Html;
-use yii\bootstrap\ActiveForm;
-
 $title = 'Secciones';
 $this->params['breadcrumbs'][] = $title;
-
-$urlCrearSeccion = Url::to(['casas/crear-seccion-ajax']);
-
-if(Yii::$app->controller->action->id == 'crear-seccion') {
-
-    $js = <<<EOL
-    var max = $('#secciones-nombre').attr('maxlength');
-    $('#secciones-nombre').after($('<span id="quedan" class="label"></span>'));
-    mostrarNumero();
-    $('#secciones-nombre').on('input', function() {
-        mostrarNumero();
-    });
-
-    function mostrarNumero() {
-        var lon = max - $('#secciones-nombre').val().length;
-        var numero = $('#quedan');
-        numero.text(lon);
-        if (lon > 16) {
-            numero.removeClass('label-success');
-            numero.addClass('label-danger');
-        } else if (lon <=5) {
-            numero.removeClass('label-success');
-            numero.addClass('label-warning');
-        } else {
-            numero.removeClass('label-danger');
-            numero.removeClass('label-warning');
-            numero.addClass('label-success');
-        }
-    }
-
-    $('#seccion-form').on('beforeSubmit', function () {
-        $.ajax({
-            url: '$urlCrearSeccion',
-            type: 'POST',
-            data: {
-                'Secciones[nombre]': $('#seccion-form').yiiActiveForm('find', 'secciones-nombre').value
-            },
-            success: function (data) {
-                $('#menu-casa-usuario').html(data);
-            }
-        });
-        return false;
-    });
-EOL;
-$this->registerJs($js);
-}
 
 ?>
 <div class="panel panel-primary panel-principal">
@@ -73,36 +23,9 @@ $this->registerJs($js);
             </div>
         </div>
         <hr>
-        <div class="row">
-            <div class="col-md-3 text-center">
-                <img src="/imagenes/seccion.png" alt="">
-            </div>
-            <div class="col-md-9">
-                <?php $form = ActiveForm::begin([
-                    'id' => 'seccion-form',
+            <?= $this->render('_input-seccion', [
+                    'model' => $model,
                 ]);
-                ?>
-                <?= $form->field($model, 'nombre', [
-                    'enableAjaxValidation' => true,
-                    'validateOnChange' => false,
-                    'validateOnBlur' => false,
-                ])
-                    ->textInput([
-                        'maxlength' => 20,
-                        'style'=>'width: 35%; display: inline; margin-right: 10px;',
-                    ])
-                    ->label('Nombre de la sección', [
-                        'style' => 'display: block',
-                        ]) ?>
-                <div class="form-group">
-                    <?= Html::submitButton('Añadir', [
-                        'class' => 'btn btn-success',
-                        'id' => 'guardar-button'
-                    ]) ?>
-                </div>
-                </div>
-            </div>
-        </div>
-        <?php ActiveForm::end(); ?>
+            ?>
     </div>
 </div>
