+++ frontend/web/css/casa.css
@@ -7,7 +7,7 @@
 
 }
 
-a.collapse-toggle {
+.panel-seccion a {
     text-decoration: none;
 }
